split up UI stuff into seperate files
