+++ Index/Include/Include_IndexUI/index_ui_impl/ui/CMakeLists.txt
@@ -1,5 +1,9 @@
 target_sources(IndexUIInclude PRIVATE
         include.cc
         ui.cc
+        
+        element/element.cc
+        mapper/mapper.cc
+
         macros/index_ui_macros.cc
         macros/index_ui_macros_end.cc)
