+++ Index/Include/Include_IndexUI/index_ui_impl/ui/include.cc
@@ -1,1 +1,6 @@
+#pragma once
+
 #include "ui.cc"
+
+#include "element/element.cc"
+#include "mapper/mapper.cc"
